refactor(core): #1590 split stream_processor.py into sub-handler modules
Extract 7+ event handlers and 4 close helpers from stream_processor.py (550 lines)
into 3 focused modules:
- stream_close.py: StreamCloseHandler (close_reasoning, truncation, exception)
- stream_text.py: StreamTextHandler + minting helpers (text, reasoning)
- stream_tool.py: StreamToolHandler + sanitization (tool lifecycle)

StreamProcessor becomes a thin orchestrator (~343 lines) that instantiates
handlers and delegates. Zero functional change — pure reorganization.

All 258 streaming tests pass, import-linter passes, ruff passes.
