+++ src/lyra/core/processors/__init__.py
@@ -8,7 +8,14 @@ changes needed.
 
 from . import _scraping, explain, search, summarize, vault_add
 from .processor_registry import BaseProcessor, ProcessorRegistry, registry
+from .stream_close import StreamCloseHandler
 from .stream_processor import StreamProcessor
+from .stream_text import (
+    StreamTextHandler,
+    _mint_reasoning_block_id,
+    _mint_text_block_id,
+)
+from .stream_tool import StreamToolHandler, _sanitize_tool_result_content
 
 __all__ = [
     "_scraping",
@@ -18,6 +25,12 @@ __all__ = [
     "vault_add",
     "BaseProcessor",
     "ProcessorRegistry",
+    "StreamCloseHandler",
     "StreamProcessor",
+    "StreamTextHandler",
+    "StreamToolHandler",
+    "_mint_text_block_id",
+    "_mint_reasoning_block_id",
+    "_sanitize_tool_result_content",
     "registry",
 ]
